consolidating normalize_registry_url
